Split modes into separate effects
Not all terminals support all modes, so split into separate effects to
allow terminals to specify the modes they support
