+++ core/shared/src/main/scala/terminus/AlternateScreenMode.scala
@@ -16,24 +16,7 @@
 
 package terminus
 
-/** Changes various modes that the terminal can be in. */
-trait Mode {
-
-  /** Run the given terminal program `f` in raw mode, which means that the
-    * program can read user input a character at a time. In canonical mode,
-    * which is the default, user input is only available a line at a time.
-    */
-  def raw[F <: effect.Effect, A](f: F ?=> A): (F & effect.Mode[F]) ?=> A =
-    effect ?=> effect.raw(f)
-
-  /** Run the given terminal program `f` in application mode, which changes the
-    * input sent to the program when arrow keys are pressed. See
-    * https://invisible-island.net/xterm/xterm.faq.html#xterm_arrows
-    */
-  def application[F <: effect.Effect, A](
-      f: F ?=> A
-  ): (F & effect.Mode[F]) ?=> A =
-    effect ?=> effect.application(f)
+trait AlternateScreenMode {
 
   /** Run the given terminal program `f` in alternate screen mode, which means
     * that whatever is displayed by `f` will not been shown when the program
@@ -41,6 +24,6 @@ trait Mode {
     */
   def alternateScreen[F <: effect.Effect, A](
       f: F ?=> A
-  ): (F & effect.Mode[F]) ?=> A =
+  ): (F & effect.AlternateScreenMode[F]) ?=> A =
     effect ?=> effect.alternateScreen(f)
 }
